requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.2.5
django-admin-persian-fonts==0.2
django-app-namespace-template-loader==0.4.1
django-blog-zinnia==0.20
django-contrib-comments==2.2.0
django-haveli-admin==0.1.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.14.0
django-tagging==0.5.0
django-xmlrpc==0.1.8
idna==3.4
jdatetime==4.1.0
mots-vides==2015.5.11
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyparsing==3.1.0
pytz==2023.3
regex==2023.6.3
requests==2.28.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.13
zinnia-theme-bootstrap-4==0.6
jsmin==3.0.1